# dashboard/docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Browser Automation SaaS Public API
  version: 1.0.0
servers:
  - url: https://your-dashboard.example/api/v1
security:
  - bearerApiKey: []
paths:
...
  /agents/{id}:
    get:
      summary: Get an owned Agent
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        {
          '200': { $ref: '#/components/responses/Success' },
          '404': { $ref: '#/components/responses/Error' },
        }
...
components:
  securitySchemes:
    bearerApiKey:
      { type: http, scheme: bearer, bearerFormat: bua_live_placeholder }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
  responses:
    Success:
      description: Successful private response
      content:
        {
          application/json:
            {
              schema:
                { type: object, required: [data], properties: { data: {} } },
            },
        }
    Error:
      description: Safe public error
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  required: [error],
                  properties:
                    {
                      error:
                        {
                          type: object,
                          required: [code, message],
                          properties:
                            {
                              code: { type: string },
                              message: { type: string },
                            },
                        },
                    },
                },
            },
        }